Allow deleting multiple files at once

+++ cmd/delete.go
@@ -1,16 +1,22 @@
 package main
 
 import (
+	"errors"
 	"fmt"
 
 	"github.com/urfave/cli/v2"
 )
 
 func DeleteAction(c *cli.Context) error {
-	id := c.String("id")
-	err := svc.DeleteFile(id)
-	if err != nil {
+	if c.NArg() < 1 {
+		return errors.New("Missing FILE_ID")
+	}
+	for _, id := range c.Args().Slice() {
+		err := svc.DeleteFile(id)
+		if err != nil {
+			return err
+		}
 		fmt.Fprintf(c.App.ErrWriter, "Deleted file %s\n", id)
 	}
-	return err
+	return nil
 }
